Renamed module "internal" and "tf2/internal" to "protocol"
