+++ auth_events.go
@@ -1,7 +1,7 @@
 package steam
 
 import (
-	. "github.com/Philipp15b/go-steam/internal/steamlang"
+	. "github.com/Philipp15b/go-steam/protocol/steamlang"
 	. "github.com/Philipp15b/go-steam/steamid"
 )
