+++ generator/GoSteamLanguageGenerator/GoGen.cs
@@ -59,7 +59,7 @@ namespace GoSteamLanguageGenerator
 			sb.AppendLine("    \"github.com/golang/protobuf/proto\"");
 			sb.AppendLine("    \"github.com/Philipp15b/go-steam/steamid\"");
 			sb.AppendLine("    \"github.com/Philipp15b/go-steam/rwu\"");
-			sb.AppendLine("   . \"github.com/Philipp15b/go-steam/internal/protobuf\"");
+			sb.AppendLine("   . \"github.com/Philipp15b/go-steam/protocol/protobuf\"");
 			sb.AppendLine(")");
 			sb.AppendLine();
