+++ protocol/gamecoordinator/msg.go
@@ -1,9 +1,9 @@
 package gamecoordinator
 
 import (
-	. "github.com/Philipp15b/go-steam/internal"
-	. "github.com/Philipp15b/go-steam/internal/protobuf"
-	. "github.com/Philipp15b/go-steam/internal/steamlang"
+	. "github.com/Philipp15b/go-steam/protocol"
+	. "github.com/Philipp15b/go-steam/protocol/protobuf"
+	. "github.com/Philipp15b/go-steam/protocol/steamlang"
 	"github.com/golang/protobuf/proto"
 	"io"
 )
