+++ protocol/internal.go
@@ -1,7 +1,7 @@
-package internal
+package protocol
 
 import (
-	. "github.com/Philipp15b/go-steam/internal/steamlang"
+	. "github.com/Philipp15b/go-steam/protocol/steamlang"
 	"io"
 )
