+++ protocol/steamlang/messages.go
@@ -5,7 +5,7 @@ package steamlang
 
 import (
 	"encoding/binary"
-	. "github.com/Philipp15b/go-steam/internal/protobuf"
+	. "github.com/Philipp15b/go-steam/protocol/protobuf"
 	"github.com/Philipp15b/go-steam/rwu"
 	"github.com/Philipp15b/go-steam/steamid"
 	"github.com/golang/protobuf/proto"
